backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0

# LLM Providers
google-generativeai==0.8.0
anthropic==0.34.0

# RAG Pipeline
rank-bm25==0.2.2
chromadb==0.5.0
sentence-transformers==3.0.0

# NER
spacy==3.7.5

# PDF Generation
reportlab==4.2.0
jinja2==3.1.4

# Database
sqlalchemy==2.0.32
aiosqlite==0.20.0

# CORS
python-multipart==0.0.9
